Allow the changing of the clients options file name by a config.h define (OPTION_FILE_NAME)
Enabeld the options file saving in the Mui Client.

Changed the "double res" declaration in tech_turns_to_advance()
of common/tech.c to "int res".

[[originally from CVS; cvs2svn created svn r2264]]

+++ common/tech.c
@@ -219,7 +219,7 @@ int find_tech_by_flag( int index, int flag )
 **************************************************************************/
 int tech_turns_to_advance(struct player *pplayer)
 {
-  double res = 0;
+  int res = 0;
 
   city_list_iterate(pplayer->cities, pcity)
     res += pcity->science_total;
